requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
alembic==1.13.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
WTForms==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
pdfkit==1.0.0
pytesseract==0.3.10
Pillow>=10.4.0
# The following packages are optional for enhanced OCR but not required:
# opencv-python==4.8.0.76
# numpy==1.25.2 